Added spinner for PR, working towards RaceResultsActivity that shows a list of results for a given racer

+++ app/src/main/java/com/nglessner/timetrial/Race.java
@@ -8,6 +8,6 @@ public class Race {
     public int RiderId;
     public int StartTime;
     public int EndTime;
-    public int Time;
+    public long Time;
     public int MPH;
 }
